requirements.txt
# Core Dependencies
numpy==1.24.3
sounddevice==0.4.6
python-dotenv==1.0.0
PyYAML==6.0.1
pydantic==2.5.2
schedule==1.2.1
parsedatetime==2.6
python-dateutil==2.8.2

# Speech Recognition
faster-whisper==0.10.0
openwakeword==0.5.0
pyaudio==0.2.13

# Text to Speech
pyttsx3==2.90

# NLP and ML
transformers==4.31.0
torch==2.0.1
torchaudio==2.0.2
sentence-transformers==2.2.2

# Database and Storage
chromadb==0.4.22
SQLAlchemy==2.0.20

# UI and Formatting
rich==13.7.0
click==8.1.7

# API Integration (Optional)
openai==1.3.0
google-generativeai==0.3.0
elevenlabs==0.2.26
serpapi==0.1.5
google-auth-oauthlib==0.4.6
google-auth-httplib2==0.1.0
google-api-python-client==2.86.0
requests==2.31.0
aiohttp==3.8.5
pydantic==2.1.1
spotipy==2.23.0
python-vlc==3.0.18122
PyYAML==6.0.1

# Development Tools
black==23.7.0
isort==5.12.0
mypy==1.5.1

# Testing
pytest==7.4.3
pytest-asyncio==0.23.2
pytest-cov==4.1.0

# Documentation
mkdocs==1.5.3
mkdocs-material==9.4.2
